refactor: migrate slideshow pure functions to party-display-core
